FIX: 'sum_array_values' in now unsigned long long

+++ src/pureCython/mean/mean.pyx
@@ -5,14 +5,14 @@ from cython cimport Py_ssize_t
 
 @cython.boundscheck(False)
 @cython.wraparound(False)
-cdef double get_mean_value_cython(np.ndarray[unsigned long, ndim=1] array):
+cdef long double get_mean_value_cython(np.ndarray[unsigned long, ndim=1] array):
     cdef const unsigned long *data_ptr = <unsigned long *> array.data
     cdef Py_ssize_t array_size = array.size
-    cdef unsigned long sum_array_values = 0
+    cdef unsigned long long sum_array_values = 0
 
     for i in range(array_size):
         sum_array_values += data_ptr[i]
-
+    
     return sum_array_values / array_size
 
 def get_mean_value_py_wrapper(np.ndarray[unsigned long, ndim=1] array):
